# pkg/api/swagger-docs/swagger.yaml
basePath: /
definitions:
...
  github_com_SENERGY-Platform_mgw-module-manager_lib_models.AuxiliaryDeployment:
    properties:
      configs:
        additionalProperties:
          type: string
        description: '{varName:value}'
        type: object
      container:
        $ref: '#/definitions/models.Container'
      created:
        type: string
      deployment_id:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      image:
        type: string
      labels:
        additionalProperties:
          type: string
        description: '{name:value}'
        type: object
      name:
        type: string
      recreate:
        type: boolean
      reference:
        type: string
      run_config:
        $ref: '#/definitions/models.AuxiliaryDeploymentRunConfig'
      updated:
        type: string
      volumes:
        additionalProperties:
          type: string
        description: '{mntPoint:ref}'
        type: object
    type: object
...
  models.AuxiliaryDeploymentReduced:
    properties:
      container:
        $ref: '#/definitions/models.Container'
      created:
        type: string
      deployment_id:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      image:
        type: string
      name:
        type: string
      recreate:
        type: boolean
      reference:
        type: string
      run_config:
        $ref: '#/definitions/models.AuxiliaryDeploymentRunConfig'
      updated:
        type: string
    type: object
...
  models.AuxiliaryDeploymentVolume:
    properties:
      deployment_id:
        type: string
      id:
        type: string
      name:
        type: string
      reference:
        type: string
    type: object
...
  models.AuxiliaryDeploymentVolumeWithMounts:
    properties:
      deployment_id:
        type: string
      id:
        type: string
      mounted_by:
        items:
          type: string
        type: array
      name:
        type: string
      reference:
        type: string
    type: object
...
  models.Job:
    properties:
      description:
        type: string
      end:
        type: string
      id:
        type: string
      start:
        type: string
    type: object
...
info:
  contact: {}
  description: Central service of the MGW ("multi gateway") edge-gateway stack.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: module-manager API
  version: dev
paths:
...
  /deployments/{DEP_ID}/auxiliary/deployments:
    get:
      description: Get the auxiliary deployments of a deployment.
      parameters:
      - description: deployment ID
        in: path
        name: DEP_ID
        required: true
        type: string
      - collectionFormat: csv
        description: filter by auxiliary deployment IDs
        in: query
        items:
          type: string
        name: ids
        type: array
      - collectionFormat: csv
        description: 'filter by labels, item format: key|value'
        in: query
        items:
          type: string
        name: labels
        type: array
      - description: filter by container image
        in: query
        name: image
        type: string
      - description: 'filter by enabled state, values: -1,0,1'
        in: query
        name: enabled
        type: integer
      - description: 'filter by recreate state, values: -1,0,1'
        in: query
        name: recreate
        type: integer
      - description: filter by container state
        in: query
        name: state
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: auxiliary deployments by ID
          schema:
            additionalProperties:
              $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-manager_lib_models.AuxiliaryDeployment'
            type: object
        "400":
          description: error message
          schema:
            type: string
        "404":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      summary: Get auxiliary deployments
      tags:
      - auxiliary-deployments
  /deployments/{DEP_ID}/auxiliary/deployments-reduced:
    get:
      description: Get the auxiliary deployments of a deployment with a reduced set
        of fields.
      parameters:
      - description: deployment ID
        in: path
        name: DEP_ID
        required: true
        type: string
      - collectionFormat: csv
        description: filter by auxiliary deployment IDs
        in: query
        items:
          type: string
        name: ids
        type: array
      - collectionFormat: csv
        description: 'filter by labels, item format: key|value'
        in: query
        items:
          type: string
        name: labels
        type: array
      - description: filter by container image
        in: query
        name: image
        type: string
      - description: 'filter by enabled state, values: -1,0,1'
        in: query
        name: enabled
        type: integer
      - description: 'filter by recreate state, values: -1,0,1'
        in: query
        name: recreate
        type: integer
      - description: filter by container state
        in: query
        name: state
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: reduced auxiliary deployments by ID
          schema:
            additionalProperties:
              $ref: '#/definitions/models.AuxiliaryDeploymentReduced'
            type: object
        "400":
          description: error message
          schema:
            type: string
        "404":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      summary: Get reduced auxiliary deployments
      tags:
      - auxiliary-deployments
  /deployments/{DEP_ID}/auxiliary/deployments/{AUX_DEP_ID}:
    get:
      description: Get a single auxiliary deployment of a deployment.
      parameters:
      - description: deployment ID
        in: path
        name: DEP_ID
        required: true
        type: string
      - description: auxiliary deployment ID
        in: path
        name: AUX_DEP_ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: auxiliary deployment
          schema:
            $ref: '#/definitions/github_com_SENERGY-Platform_mgw-module-manager_lib_models.AuxiliaryDeployment'
        "404":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      summary: Get auxiliary deployment
      tags:
      - auxiliary-deployments
  /deployments/{DEP_ID}/auxiliary/volumes:
    get:
      description: Get the auxiliary deployment volumes of a deployment.
      parameters:
      - description: deployment ID
        in: path
        name: DEP_ID
        required: true
        type: string
      - collectionFormat: csv
        description: filter by volume references
        in: query
        items:
          type: string
        name: references
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: auxiliary deployment volumes by reference
          schema:
            additionalProperties:
              $ref: '#/definitions/models.AuxiliaryDeploymentVolume'
            type: object
        "400":
          description: error message
          schema:
            type: string
        "404":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      summary: Get auxiliary deployment volumes
      tags:
      - auxiliary-deployments
  /deployments/{DEP_ID}/auxiliary/volumes-with-mounts:
    get:
      description: Get the auxiliary deployment volumes of a deployment including
        the auxiliary deployments they are mounted in.
      parameters:
      - description: deployment ID
        in: path
        name: DEP_ID
        required: true
        type: string
      - collectionFormat: csv
        description: filter by volume references
        in: query
        items:
          type: string
        name: references
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: auxiliary deployment volumes with mounts by reference
          schema:
            additionalProperties:
              $ref: '#/definitions/models.AuxiliaryDeploymentVolumeWithMounts'
            type: object
        "400":
          description: error message
          schema:
            type: string
        "404":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      summary: Get auxiliary deployment volumes with mounts
      tags:
      - auxiliary-deployments
...
  /restricted/deployments/{DEP_ID}/auxiliary/deployments-disable:
    post:
      consumes:
      - application/json
      description: Disable the auxiliary deployments of a deployment matching the
        given filter.
      parameters:
      - description: deployment ID
        in: path
        name: DEP_ID
        required: true
        type: string
      - description: auxiliary deployments filter
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AuxiliaryDeploymentsFilterWithState'
      produces:
      - application/json
      responses:
        "200":
          description: IDs of the disabled auxiliary deployments
          schema:
            items:
              type: string
            type: array
        "400":
          description: error message
          schema:
            type: string
        "404":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      summary: Disable auxiliary deployments
      tags:
      - auxiliary-deployments
  /restricted/deployments/{DEP_ID}/auxiliary/deployments-enable:
    post:
      consumes:
      - application/json
      description: Enable the auxiliary deployments of a deployment matching the given
        filter.
      parameters:
      - description: deployment ID
        in: path
        name: DEP_ID
        required: true
        type: string
      - description: auxiliary deployments filter
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AuxiliaryDeploymentsFilterWithState'
      produces:
      - application/json
      responses:
        "200":
          description: IDs of the enabled auxiliary deployments
          schema:
            items:
              type: string
            type: array
        "400":
          description: error message
          schema:
            type: string
        "404":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      summary: Enable auxiliary deployments
      tags:
      - auxiliary-deployments
  /restricted/deployments/{DEP_ID}/auxiliary/deployments-recreate:
    post:
      consumes:
      - application/json
      description: Recreate the auxiliary deployments of a deployment matching the
        given filter as a job.
      parameters:
      - description: deployment ID
        in: path
        name: DEP_ID
        required: true
        type: string
      - description: auxiliary deployments filter
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AuxiliaryDeploymentsFilterWithState'
      produces:
      - application/json
      responses:
        "200":
          description: job
          schema:
            $ref: '#/definitions/models.Job'
        "400":
          description: error message
          schema:
            type: string
        "404":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
        "503":
          description: error message
          schema:
            type: string
      summary: Recreate auxiliary deployments
      tags:
      - auxiliary-deployments
...
  /restricted/deployments/{DEP_ID}/auxiliary/volumes-with-mounts:
    get:
      description: Get the auxiliary deployment volumes of a deployment including
        the auxiliary deployments they are mounted in.
      parameters:
      - description: deployment ID
        in: path
        name: DEP_ID
        required: true
        type: string
      - collectionFormat: csv
        description: filter by volume references
        in: query
        items:
          type: string
        name: references
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: auxiliary deployment volumes with mounts by reference
          schema:
            additionalProperties:
              $ref: '#/definitions/models.AuxiliaryDeploymentVolumeWithMounts'
            type: object
        "400":
          description: error message
          schema:
            type: string
        "404":
          description: error message
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      summary: Get auxiliary deployment volumes with mounts
      tags:
      - auxiliary-deployments
...
swagger: "2.0"